+++ src/components/Radio/Radio.test.tsx
@@ -54,6 +54,23 @@ describe("Radio Button", () => {
     expect(changeHandler.callCount).to.equal(1);
   });
 
+  it("Renders with appropriate 'aria-label' attribute and value when 'showLabel' prop is set to false and 'helperText' has been passed", () => {
+    container = Enzyme.mount(
+      <Radio
+        //add other props here
+        id="test_id"
+        labelText="Hello"
+        showLabel={false}
+        helperText="The helper text."
+      ></Radio>
+    );
+    const labelText = container.prop("labelText");
+    const helperText = container.prop("helperText");
+    expect(container.find("input").prop("aria-label")).to.equal(
+      `${labelText} - ${helperText}`
+    );
+  });
+
   it("Renders with label", () => {
     container = Enzyme.mount(
       <Radio
@@ -67,6 +84,36 @@ describe("Radio Button", () => {
     expect(container.find("input").props()).not.to.have.property("aria-label");
     const radioId = container.prop("id");
     expect(container.find("label").prop("htmlFor")).to.equal(radioId);
+  });
+
+  it("Renders with visible helper text", () => {
+    container = Enzyme.mount(
+      <Radio
+        //add other props here
+        id="test_id"
+        labelText="Hello"
+        showLabel={true}
+        helperText="The helper text."
+        errorText="The error text."
+        errored={false}
+      ></Radio>
+    );
+    expect(container.find(".helper-text").text()).to.equal("The helper text.");
+  });
+
+  it("Renders with visible error text", () => {
+    container = Enzyme.mount(
+      <Radio
+        //add other props here
+        id="test_id"
+        labelText="Hello"
+        showLabel={true}
+        helperText="The helper text."
+        errorText="The error text."
+        errored={true}
+      ></Radio>
+    );
+    expect(container.find(".helper-text").text()).to.equal("The error text.");
   });
 
   it("Calls a UUID generation method if no ID is passed as a prop", () => {