+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Adds CSS vars for `font-weight`
 - Adds `displaySize` prop to `Heading`
+- Adds `helperText` and `errorText` props to `Radio`
 
 ### Changes